requirements.txt
Flask
gunicorn
reportlab
PyPDF2
python-docx
google-generativeai
python-dotenv
